Revert "Added type checking and directory structure changes"
This reverts commit 9f175e8cace96a0bfe2017872b9ce59f9a3b241e.
