+++ .gitignore
@@ -171,4 +171,3 @@ cython_debug/
 
 # pyenv
 .python-version
-pyrightconfig.json
